+++ client.go
@@ -1,23 +1,20 @@
 package sse
 
 import (
-	"context"
 	"fmt"
+	"math/rand"
 	"net/http"
 	"strings"
 	"time"
 	"unicode"
-
-	"github.com/cenkalti/backoff/v4"
 )
 
 // The ResponseValidator type defines the type of the function
 // that checks whether server responses are valid, before starting
 // to read events from them. See the Client's documentation for more info.
 //
-// Returning an error type that implements the Temporary method will
-// tell the client to reattempt the connection. This can be useful if
-// you're being rate limited (response code 429), for example.
+// These errors are considered permanent and thus if the client is configured
+// to retry on error no retry is attempted and the error is returned.
 type ResponseValidator func(*http.Response) error
 
 // The Client struct is used to initialize new connections to different servers.
@@ -29,7 +26,8 @@ type Client struct {
 	// The HTTP client to be used. Defaults to http.DefaultClient.
 	HTTPClient *http.Client
 	// A callback that's executed whenever a reconnection attempt starts.
-	OnRetry backoff.Notify
+	// It receives the error that caused the retry and the reconnection time.
+	OnRetry func(error, time.Duration)
 	// A function to check if the response from the server is valid.
 	// Defaults to a function that checks the response's status code is 200
 	// and the content type is text/event-stream.
@@ -39,16 +37,35 @@ type Client struct {
 	// Otherwise, the error will be considered permanent and no reconnections
 	// will be attempted.
 	ResponseValidator ResponseValidator
-	// The maximum number of reconnection to attempt when an error occurs.
-	// If MaxRetries is negative (-1), infinite reconnection attempts will be done.
-	// Defaults to 0 (no retries).
-	//
-	// This counter is reset if a reconnection attempt is successful.
+	// Backoff configures the backoff strategy. See the documentation of
+	// each field for more information.
+	Backoff Backoff
+}
+
+// Backoff configures the reconnection strategy of a Connection.
+type Backoff struct {
+	// The initial wait time before a reconnection is attempted.
+	// Must be >0. Defaults to 500ms.
+	InitialInterval time.Duration
+	// How much should the reconnection time grow on subsequent attempts.
+	// Must be >=1; 1 = constant interval. Defaults to 1.5.
+	Multiplier float64
+	// How much does the reconnection time vary relative to the base value.
+	// This is useful to prevent multiple clients to reconnect at the exact
+	// same time, as it makes the wait times distinct.
+	// Must be in range (0, 1); -1 = no randomization. Defaults to 0.5.
+	Jitter float64
+	// How much can the wait time grow.
+	// If <=0 = the wait time can infinitely grow. Defaults to infinite growth.
+	MaxInterval time.Duration
+	// How much time can retries be attempted.
+	// For example, if this is 5 seconds, after 5 seconds the client
+	// will stop retrying.
+	// If <=0 = no limit. Defaults to no limit.
+	MaxElapsedTime time.Duration
+	// How many retries are allowed.
+	// <0 = no retries, 0 = infinite. Defaults to infinite retries.
 	MaxRetries int
-	// The initial reconnection delay. Subsequent reconnections use a longer
-	// time. This can be overridden by retry values sent by the server.
-	// Defaults to 5 seconds.
-	DefaultReconnectionTime time.Duration
 }
 
 // NewConnection initializes and configures a connection. On connect, the given
@@ -74,33 +91,6 @@ func (c *Client) NewConnection(r *http.Request) *Connection {
 	}
 
 	return conn
-}
-
-func (c *Client) newBackoff(ctx context.Context) (b backoff.BackOff, setRetry func(time.Duration)) {
-	base := backoff.NewExponentialBackOff()
-	base.InitialInterval = c.DefaultReconnectionTime
-	b = backoff.WithContext(base, ctx)
-	if c.MaxRetries >= 0 {
-		rb := backoff.WithMaxRetries(b, uint64(c.MaxRetries))
-		return rb, func(d time.Duration) {
-			base.InitialInterval = d
-			rb.Reset()
-		}
-	}
-	return b, func(d time.Duration) {
-		base.InitialInterval = d
-		b.Reset()
-	}
-}
-
-func contentType(header string) string {
-	cts := strings.FieldsFunc(header, func(r rune) bool {
-		return unicode.IsSpace(r) || r == ';' || r == ','
-	})
-	if len(cts) == 0 {
-		return ""
-	}
-	return strings.ToLower(cts[0])
 }
 
 // DefaultValidator is the default client response validation function. As per the spec,
@@ -129,9 +119,13 @@ var NoopValidator ResponseValidator = func(_ *http.Response) error {
 // DefaultClient is the client that is used when creating a new connection using the NewConnection function.
 // Unset properties on new clients are replaced with the ones set for the default client.
 var DefaultClient = &Client{
-	HTTPClient:              http.DefaultClient,
-	DefaultReconnectionTime: time.Second * 5,
-	ResponseValidator:       DefaultValidator,
+	HTTPClient:        http.DefaultClient,
+	ResponseValidator: DefaultValidator,
+	Backoff: Backoff{
+		InitialInterval: time.Millisecond * 500,
+		Multiplier:      1.5,
+		Jitter:          0.5,
+	},
 }
 
 // NewConnection creates a connection using the default client.
@@ -143,13 +137,93 @@ func mergeDefaults(c *Client) {
 	if c.HTTPClient == nil {
 		c.HTTPClient = DefaultClient.HTTPClient
 	}
-	if c.MaxRetries == 0 {
-		c.MaxRetries = DefaultClient.MaxRetries
+	if c.Backoff.InitialInterval <= 0 {
+		c.Backoff.InitialInterval = DefaultClient.Backoff.InitialInterval
 	}
-	if c.DefaultReconnectionTime <= 0 {
-		c.DefaultReconnectionTime = DefaultClient.DefaultReconnectionTime
+	if c.Backoff.Multiplier < 1 {
+		c.Backoff.Multiplier = DefaultClient.Backoff.Multiplier
+	}
+	if c.Backoff.Jitter <= 0 || c.Backoff.Jitter >= 1 {
+		c.Backoff.Jitter = DefaultClient.Backoff.Jitter
 	}
 	if c.ResponseValidator == nil {
 		c.ResponseValidator = DefaultClient.ResponseValidator
 	}
+}
+
+func contentType(header string) string {
+	cts := strings.FieldsFunc(header, func(r rune) bool {
+		return unicode.IsSpace(r) || r == ';' || r == ','
+	})
+	if len(cts) == 0 {
+		return ""
+	}
+	return strings.ToLower(cts[0])
+}
+
+type backoffController struct {
+	start      time.Time
+	rng        *rand.Rand
+	b          *Backoff
+	interval   time.Duration
+	numRetries int
+}
+
+func (b *Backoff) new() backoffController {
+	now := time.Now()
+	return backoffController{
+		start:      now,
+		rng:        rand.New(rand.NewSource(now.UnixNano())),
+		b:          b,
+		interval:   b.InitialInterval,
+		numRetries: 0,
+	}
+}
+
+// reset the backoff to the initial state, i.e. as if no retries have occurred.
+// If newInterval is greater than 0, the initial interval is changed to it.
+func (c *backoffController) reset(newInterval time.Duration) {
+	if newInterval > 0 {
+		c.interval = newInterval
+	} else {
+		c.interval = c.b.InitialInterval
+	}
+	c.numRetries = 0
+	c.start = time.Now()
+}
+
+func (c *backoffController) next() (interval time.Duration, shouldRetry bool) {
+	if c.b.MaxRetries < 0 || (c.b.MaxRetries > 0 && c.numRetries == c.b.MaxRetries) {
+		return 0, false
+	}
+
+	c.numRetries++
+	elapsed := time.Since(c.start)
+	next := nextInterval(c.b.Jitter, c.rng, c.interval)
+	c.interval = growInterval(c.interval, c.b.MaxInterval, c.b.Multiplier)
+
+	if c.b.MaxElapsedTime > 0 && elapsed+next > c.b.MaxElapsedTime {
+		return 0, false
+	}
+
+	return next, true
+}
+
+func nextInterval(jitter float64, rng *rand.Rand, current time.Duration) time.Duration {
+	if jitter == -1 {
+		return current
+	}
+
+	delta := jitter * float64(current)
+	minInterval := float64(current) - delta
+	maxInterval := float64(current) + delta
+
+	return time.Duration(minInterval + (rng.Float64() * (maxInterval - minInterval + 1)))
+}
+
+func growInterval(current, maxInterval time.Duration, mul float64) time.Duration {
+	if maxInterval > 0 && float64(current) >= float64(maxInterval)/mul {
+		return maxInterval
+	}
+	return time.Duration(float64(current) * mul)
 }